backend/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
django-cors-headers==3.5.0
django==3.1.4; python_version >= '3.6'
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1; python_version >= '3.5'
